# api/gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Tekton Hub","description":"HTTP services for managing Tekton Hub","version":"1.0"},"host":"localhost:8000","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/categories":{"get":{"tags":["category"],"summary":"All category","description":"Get all Categories with their tags sorted by name","operationId":"category#All","responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"$ref":"#/definitions/categoryResponse"}}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/CategoryAllInternalErrorResponseBody"}}},"schemes":["http"]}},"/resource/{resourceId}/versions":{"get":{"tags":["resourceversions"],"summary":"Resourceversionsdetail resourceversions","description":"Get all versions information of a resource by resourceId","operationId":"resourceversions#Resourceversionsdetail","parameters":[{"name":"resourceId","in":"path","description":"id of resource","required":true,"type":"integer"}],"responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"$ref":"#/definitions/ResourceversionResponse"}}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/ResourceversionsResourceversionsdetailInternalErrorResponseBody"}}},"schemes":["http"]}},"/swagger/swagger.json":{"get":{"tags":["swagger"],"summary":"Download gen/http/openapi.json","description":"JSON document containing the API swagger definition","operationId":"swagger#/swagger/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"CatalogResponse":{"title":"CatalogResponse","type":"object","properties":{"id":{"type":"integer","description":"ID is the unique id of the category","example":1,"format":"int64"},"type":{"type":"string","description":"Type of support tier","example":"official"}},"example":{"id":1,"type":"official"},"required":["id","type"]},"CategoryAllInternalErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"All_internal-error_Response_Body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"ResourceVersionInfoResponse":{"title":"ResourceVersionInfoResponse","type":"object","properties":{"description":{"type":"string","description":"description of perticular version","example":"This Task builds source into a container image"},"displayName":{"type":"string","description":"displayname of perticular version","example":"git clone"},"id":{"type":"integer","description":"resource version id","example":1,"format":"int64"},"rawUrl":{"type":"string","description":"github raw url of perticular veriosn of a resource","example":"https://raw.github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/"},"updatedAt":{"type":"string","description":"Date when resource was last updated","example":"2020-03-26T14:09:08.19599+05:30"},"version":{"type":"string","description":"version name","example":"1.0"},"webUrl":{"type":"string","description":"github web url of perticular veriosn of a resource","example":"https://github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/"}},"example":{"description":"This Task builds source into a container image","displayName":"git clone","id":1,"rawUrl":"https://raw.github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/","updatedAt":"2020-03-26T14:09:08.19599+05:30","version":"1.0","webUrl":"https://github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/"},"required":["id","version","description","displayName","rawUrl","webUrl","updatedAt"]},"ResourceversionResponse":{"title":"Mediatype identifier: resourceversion; view=default","type":"object","properties":{"catalog":{"$ref":"#/definitions/CatalogResponse"},"id":{"type":"integer","description":"ID is the unique id of the resource","example":1,"format":"int64"},"name":{"type":"string","description":"Name of the resource","example":"golang"},"rating":{"type":"integer","description":"Rating of resource","example":5,"format":"int64"},"tags":{"type":"array","items":{"$ref":"#/definitions/TagResponse"},"description":"Tags related to resources","example":[{"id":1,"name":"cli"},{"id":1,"name":"cli"}]},"type":{"type":"string","description":"Type of resource","example":"task"},"versions":{"type":"array","items":{"$ref":"#/definitions/ResourceVersionInfoResponse"},"description":" versions information of a resource","example":[{"description":"This Task builds source into a container image","displayName":"git clone","id":1,"rawUrl":"https://raw.github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/","updatedAt":"2020-03-26T14:09:08.19599+05:30","version":"1.0","webUrl":"https://github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/"},{"description":"This Task builds source into a container image","displayName":"git clone","id":1,"rawUrl":"https://raw.github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/","updatedAt":"2020-03-26T14:09:08.19599+05:30","version":"1.0","webUrl":"https://github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/"},{"description":"This Task builds source into a container image","displayName":"git clone","id":1,"rawUrl":"https://raw.github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/","updatedAt":"2020-03-26T14:09:08.19599+05:30","version":"1.0","webUrl":"https://github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/"}]}},"description":"ResourceversionResponse result type (default view)","example":{"catalog":{"id":1,"type":"official"},"id":1,"name":"golang","rating":5,"tags":[{"id":1,"name":"cli"},{"id":1,"name":"cli"},{"id":1,"name":"cli"},{"id":1,"name":"cli"}],"type":"task","versions":[{"description":"This Task builds source into a container image","displayName":"git clone","id":1,"rawUrl":"https://raw.github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/","updatedAt":"2020-03-26T14:09:08.19599+05:30","version":"1.0","webUrl":"https://github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/"},{"description":"This Task builds source into a container image","displayName":"git clone","id":1,"rawUrl":"https://raw.github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/","updatedAt":"2020-03-26T14:09:08.19599+05:30","version":"1.0","webUrl":"https://github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/"},{"description":"This Task builds source into a container image","displayName":"git clone","id":1,"rawUrl":"https://raw.github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/","updatedAt":"2020-03-26T14:09:08.19599+05:30","version":"1.0","webUrl":"https://github.com/Pipelines-Marketplace/catalog/tree/master/task/kaniko/0.1/"}]},"required":["id","name","catalog","type","rating","tags","versions"]},"ResourceversionsResourceversionsdetailInternalErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Resourceversionsdetail_internal-error_Response_Body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"TagResponse":{"title":"TagResponse","type":"object","properties":{"id":{"type":"integer","description":"Id is the unique id of tags","example":1,"format":"int64"},"name":{"type":"string","description":"name of tag","example":"cli"}},"example":{"id":1,"name":"cli"},"required":["id","name"]},"categoryResponse":{"title":"categoryResponse","type":"object","properties":{"id":{"type":"integer","description":"unique id of category","example":1,"format":"int64"},"name":{"type":"string","description":"name of category","example":"Image-build"},"tags":{"type":"array","items":{"$ref":"#/definitions/TagResponse"},"description":"list of tag associated with category","example":[{"id":1,"name":"cli"},{"id":1,"name":"cli"},{"id":1,"name":"cli"}]}},"example":{"id":1,"name":"Image-build","tags":[{"id":1,"name":"cli"},{"id":1,"name":"cli"},{"id":1,"name":"cli"},{"id":1,"name":"cli"}]},"required":["id","name","tags"]}}}